feat(management/permission): add permission in pages
